refactor: moved deprecations from index
