+++ packages/core/src/features/index.ts
@@ -9,7 +9,7 @@ export type { ImportSymbol, Imported } from './st-import';
 export * as STGlobal from './st-global';
 
 export * as STVar from './st-var';
-export type { VarSymbol } from './st-var';
+export type { VarSymbol, ComputedStVar, FlatComputedStVar } from './st-var';
 
 export * as STMixin from './st-mixin';
 export type { RefedMixin, MixinValue } from './st-mixin';
